requirements.txt
numpy==1.20.3
Flask-SQLAlchemy==2.4.4
scikit-image==0.18.1
scikit-learn==0.24.0
opencv-python-headless==4.5.3.56
scipy==1.6.0
requests==2.25.1
SQLAlchemy==1.3.22
protobuf==3.20.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1
torchvision==0.9.1
torchaudio==0.8.1
torchsummary==1.5.1
tensorboard==2.4.1
torchinfo==1.6.6
Flask-Restless==0.17.0
Flask==1.1.2
umap-learn==0.5.1
numba==0.54.0
llvmlite==0.37.0
Pillow==8.1.2
albumentations==0.4.3
dill==0.3.3
Shapely==1.7.1
tables==3.6.1
tqdm==4.59.0
matplotlib==3.3.4
ttach==0.0.2
config==0.4.2
flask-restx==0.3.0
waitress==2.0.0
Flask-Admin==1.5.8
Flask-Caching==1.10.1
itsdangerous==1.1.0
Jinja2==3.0.3
Werkzeug==2.0.3
thefuzz==0.19.0
